+++ src/main/java/de/fanta/cubeside/mixin/LogicalZoomMixin.java
@@ -16,9 +16,9 @@ import net.minecraft.client.render.GameRenderer;
 public class LogicalZoomMixin {
     //Code from LogicalGeekBoy -> https://github.com/LogicalGeekBoy/logical_zoom
     @Inject(method = "getFov(Lnet/minecraft/client/render/Camera;FZ)F", at = @At("RETURN"), cancellable = true)
-    public void getZoomLevel(CallbackInfoReturnable<Double> callbackInfo) {
+    public void getZoomLevel(CallbackInfoReturnable<Float> callbackInfo) {
         if (LogicalZoom.isZooming()) {
-            double fov = callbackInfo.getReturnValue();
+            float fov = callbackInfo.getReturnValue();
             callbackInfo.setReturnValue(fov * LogicalZoom.zoomLevel);
         }
